fix: ech for hellogolang
